+++ HumanResource/database/migrations/2024_09_26_084947_create_departments_table.php
@@ -12,13 +12,15 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('departments', function (Blueprint $table) {
-            $table->smallInteger('DepartmentID')->primary(); // Khóa chính
+            $table->bigIncrements('DepartmentID')->primary(); // Khóa chính
             $table->longText('Name');
             $table->longText('GroupName');
-            $table->timestamp('ModifiedDate')->nullable(); // Sử dụng timestamp cho ModifiedDate
-            $table->timestamps();
+            $table->timestamp('ModifiedDate')->nullable(); // Ngày cập nhật cuối
+            $table->timestamps(); // Tạo cột created_at và updated_at
         });
+         
     }
+
 
     /**
      * Reverse the migrations.
